Git: middle patch may be empty in merge
When there is no difference between two (distinct) merge heads, the middle
patch is empty.

In the following example, the heads h1 and h2 are distinct
(isNothingToUpdate would return false). However, there is no difference
between the two states, so the middle patch is empty.

    0  (+A)  (+B)  (-B)
    *----*-----*-----* h1
          \
           *---------* h2
         (+C)      (-C)

Therefore, care must be taken when building the middle patch: the
PatchBuilder may return null.