Repository: fixed commonAncestor() bug
There was a bug in Repository.commonAncestor() when one of the two
commits is the ancestor ("fast forward" or "nothing to update"
situations). The two special cases were correctly identified in the
javadoc, but inverted in the implementation.

This bug was hidden because of another bug in Git.merge(), where the
"forward" field was (incorrectly) set to the common ancestor.